Fix: race between commands and pid appearance
Add init_done flag to fix race where a PID appears simultaneously with
the execution of a sequence of commands, for instance creating a channel
and then an event. We had to make the ust app visible while it's not
fully initialized (all its app sessions being up to date with the global
ust app sessions) because of the ust registry. Add a init_done flag to
fix this.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>